pyproject.toml
[project]
name = "src"
version = "0.1.0"
authors = [{ name = "Kuriko Iwai", email = "krik8235@gmail.com" }]
description = "sales prediction"
readme = "README.md"
requires-python = "==3.12.*"
license = {file = "LICENSE"}
keywords = ["machine learning", 'deep learning', 'time series']
classifiers = ["Programming Language :: Python"]
dependencies = [
    "apache-airflow>=3.0.6",
    "boto3>=1.39.17",
    "category-encoders>=2.8.1",
    "deltalake>=1.1.4",
    "numpy>=2.1.3",
    "pandas>=2.3.1",
    "pyspark>=4.0.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "torch>=2.9.0",
    "yfinance>=0.2.66",
]

[project.urls]
Homepage = "https://kuriko.vercel.app"
Repository = "https://github.com/krik8235/ml-sales-prediction"
Issues = "https://github.com/krik8235/ml-sales-prediction/issues"

[project.optional-dependencies]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.uv]
package = true

[tool.uv.workspace]
members = ["kuriko"]

[tool.mypy]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = [""]

[dependency-groups]
dev = [
    "evidently>=0.7.15",
    "fastparquet>=2024.11.0",
    "mlflow>=3.5.1",
    "pyarrow>=21.0.0",
    "pydeequ>=1.5.0",
    "pytest>=8.4.2",
    "s3fs>=2025.10.0",
]
